Fixed the ComboBox/ListBox
Whenever an item from the ListBox is selected, the name/password combo will be updated on the textboxes.

Plan on converting the combo box into a hidden 2nd listbox to store the password values and save every item in those listboxes to load them back again whenever the app is launched.