Added functionality
Added a Save/Load button that will save/load all the objects stored in the listBoxes.
Added some logic to prevent duplicates when adding a new item to the listBox.
Added a Remove button to remove the selected object from the listBox & a Clear button to wipe out every single object on the list.