More implementantions
Added a password field that hides the password and a checkbox that will display the password.
The client button now checks whether the user can join the client or not.
Added a Restart button to launch a new process and kill the one that was being used.
The app now prevents itself from running two processes of it at the same time (might remove this later though).
Currently trying to detach the game window with a click of a button (to be able to play the game on a bigger window, if needed) but might have to switch to something else. The detach button is simply running a new Window, for now.