Removed WindowsInput
The program no longer relies on WindowsInput to input the information. It's doing everything it was doing before but with DOM properties. It's both more efficient and accurate since it doesn't have to press Tab X times to get to the input place.

However, I'm keeping the WindowsInput package for now since it might be handy in the near future.